Use Redix instead of Exredis (#63)
What changed?
============

We change the library we use to communicate with Redis from Exredis (and
Eredis) to Redix.

Exredis seems to be going through a transition. The libraries `master`
branch has completely changed and only the [`exredis`
branch](https://github.com/artemeff/redis/tree/exredis) now contains the
library we currently use.

Moreover, exredis (and the underlying eredis) do not support using SSL
when connecting to redis. This is a hard limitation when using Redis in
places that require SSL connections (e.g. Heroku private spaces).

Redix seems to be the most popular library (when searching for "redis"
in hex), and it supports SSL. It's used by libraries like
[Exq](https://github.com/akira/exq) and
[FunWithFlags](https://github.com/tompave/fun_with_flags).

Change details
---------------

Most of the changes were confined to the `Redbird.Redis` module. It's
nice to have that abstraction so we can easily swap `Exredis.Api` usage
with `Redix`. I used `Plug.Session.REDIS` to guide what return values
are needed.

We now pass the redis connection options from the `redbird` application.
Thus, instead of relying on `ex_redis` configuration, we make it part of
redbird's configuration as `redis_options`:

```elixir
config :redbird, redis_options []
```

This has the benefit of tying the options more clearly to redbird, and
it allows applications to use different Redis stores for different
purposes: e.g. one redis store for sessions via redbird, and another for
background jobs via exq.