Add bin/release script
Adds a simple bin/release script that will change the version referenced
in the README and in `mix.exs` from the previous version to the new
version.

It then commits the release, creates a tag, and pushes to github.
Finally it runs `mix hex.publish` to publish the package.

Example usage:

```
bin/release 0.4.0 0.5.0
```