Add a test for reuse of the redis key
This test fails, exposing a bug. As things stand, Redbird accidentally
moves to a new redis key each time `Plug.Session.REDIS.put/4` is called.

This stems from calling `Redbird.Key.sign_key/2` unconditionally
(as of 299906f5), which results in signing the already-signed
user-supplied key.

We end up with a key with arbitrary levels of nested signatures.
This is bad in its own right (redbird isn't supposed to move to a new
key each `put`), but also leads to the key growing without bound,
as each new layer of signature results in a somewhat-longer key.

Eventually, this overflows the maximum length of 4096 bytes which plug
is willing to allow in a cookie, leading to a Plug.Conn.CookieOverflowError.