Avoid re-signing the redis key with each call to put/4
The function previously known as Redbird.Key.deletable?/2 is
now Redbird.Key.accessible?/2, and is used in put/4, get/3, and
delete/2.

A key is now signed exactly once, immediately after being generated,
and verified each time it is used. This fixes the bug described
in the previous commit, where the session key is re-signed
with each call to put/4, causing it to grow without bound and eventually
overflow.