Fix the broken session expiration test
The session expiration test was broken: even if you removed the
call to :timer.sleep(1000) the test would still pass.

This is because each of the two requests used a different secret,
meaning that the second conn didn't have access to the session
from the first in any case.

By using sign_conn_with/3 instead of sign_conn/2, this test can use the
same secret for both conns, ensuring that this test actually tests what
it claims to be testing.