Regenerate invalidly signed redis keys in put/4
The redis key sent in the cookie can't be trusted if it's unsigned, or
if it's signed incorrectly. We need to ensure that we don't allow
malicious users to set arbitrary redis keys.

Therefore, if the signature can't be validate we start over with a new
generated key in the same way as when the key isn't specified.